Added ability to choose widget position
this change set adds a class responsible for positioning the widget. The
options are to pick which display on a multi-display system and weather
the widget is on the left or right hand side of the display. Currently
there is no user interface for setting these values but they are being
stored in a settings file.