Added Link to and fixed size issue
When not using the same resolution monitors height and y positing
becomes a problem (my left most monitor is lower along the Y axis and
has a larger resolution then my other two monitors (my right most is in
portrate mode). This change set makes sure that the height of the widget
and Y position is based on the currently selected display.

Also added a link to the RSS artical, when the user clicks on the title
of a given RSS entry it will open the URL associated with that entry in
a new web browser.