Fixed mouse bleed through and designer issue
Turns out if you use anchors.fill: parent within your *.ui.qml file the
designer view won't work (as your trying to anchor to the designer).
This is similar to VisualStudio's deisgner view had this issue back in
the day. Also put a MouseArea in the header back so mouse clicks don't
bleed through (is an issue now that I've added links in the RSS feed
titles).